backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
mysqlclient==2.0.3
sqlparse==0.4.3
typing_extensions==4.11.0
uWSGI==2.0.22
mysqlclient==2.0.3
djangorestframework-simplejwt==5.2.2
django-redis>=5.4.0
redis==5.0.0
colorlog==6.7.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2025.12.1
django-cors-headers==3.14.0
django-csp==3.8.0
celery==5.3.6
requests
Pillow==10.2.0


google-api-python-client==2.100.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
youtube-transcript-api==0.6.2
# openai + httpx 0.28+ : httpx Client 에서 proxies 인자 제거 → OpenAI SDK와 TypeError → httpx 상한 고정
httpx>=0.27.0,<0.28
openai>=1.40.0,<2
yt-dlp==2026.2.21